webpush: Limit the number of subscriptions that can be added
The counting is a bit inefficient, because it stores all the results to an
array and only then discards it. However, this code isn't called very often
and doing it this way results in less/simpler code overall. The following
commits will make use of the added webpush_subscriptions_read().